modify : user_index 자료형 수정
user_index 자료형 Long에서 int로

+++ src/main/java/com/bringup/member/membership/domain/repository/UserMembershipRepository.java
@@ -3,6 +3,6 @@ package com.bringup.member.membership.domain.repository;
 import com.bringup.member.membership.domain.entity.UserMembership;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserMembershipRepository extends JpaRepository<UserMembership, Long> {
-    boolean existsByUserIndex(Long userIndex);
+public interface UserMembershipRepository extends JpaRepository<UserMembership, Integer> {
+    boolean existsByUserIndex(int userIndex);
 }
